Totally rejiggered how handler mappings works.
Now, many handlers can be registered and the handlers will either answer or pass on to the next handler. This makes it very flexible - a handler can act like an end point, or middleware component. And handlers can be responsible for providing path mapping schemes.

Introduce HttpControl interface which allows handlers to pass control to the next handler.