Wrap exceptions in WebbitException. This provides better documentation for exceptions.
It shows the details of the connection (if applicable) when an error occurs.

It ensures stack traces conntain "Webbit" somewhere. Since most of our code is just
instantiation of Netty code a lot of exceptions it throws don't contain any Webbit
code in the stack trace. When using Netty for a number of different networking services
it's not very obvious where an exception came from.