Fix embedded resource listing for jars and folders
Before, embedded resources were only loaded if class files were in
folders, not jars.  Testing in-IDE, or with maven does not actually
package the tests in jars; it reads them from a folder.  `make test`
*does* package them in properly jars, so picked up this failure.

Leaves requests for directories with directory listing disabled 404ing,
which is probably wrong, but no worse than it was before.

Gets rid of fragile ByteArrayInputStream=directory check.

Adds ability to specify which classpath entry should be used to serve
files from (rather than forcing the classpath entry to be whichever one
webbit itself is in) - removes the need to repackage webbit to serve
files from a jar on the classpath.

Tested with:
mvn clean test
make clean test
Running all tests in Eclipse